refactor(map): extract subject interactions from CivicMap into a hook
CivicMap mixed instance plumbing, layer wiring, interaction policy and DOM
marker rendering in one ~500-line component. Move the subjects-on-the-map
concern — the clustered subjects layer + donut pool + spiderfier, selection
highlight / fly-to, panel hover, and the unified click resolver (a subject
within 8px beats a municipality; an empty click clears) — into
useMapInteractions. The subject derivations (by-id, by-anchor, marker
options) move with it; padding, callbacks and the municipalities lookup stay
in CivicMap and are passed in as refs, so the heavy listeners still register
once and never go stale. The shared fly-to constants move to lib/map/constants.
CivicMap goes 511 -> 294 lines; behaviour-preserving (verbatim move).
